+++ packages/typespec-azure-core/src/rules/non-breaking-versioning.ts
@@ -1,4 +1,10 @@
-import { ModelProperty, Type, createRule, fileRef, paramMessage } from "@typespec/compiler";
+import {
+  type ModelProperty,
+  type Type,
+  createRule,
+  fileRef,
+  paramMessage,
+} from "@typespec/compiler";
 import {
   getAddedOnVersions,
   getMadeOptionalOn,