+++ packages/typespec-azure-resource-manager/generated-defs/Azure.ResourceManager.BaseTypes.ts
@@ -17,7 +17,7 @@ export interface BaseTypeInfo {
  * @example
  * ```typespec
  * // Agent definition and properties using the Appliance deployment model
- * model ContosoApplianceDefinition is AgentDefinitionAppliance<true, true>;
+ * model ContosoApplianceDefinition is AgentDefinitionAppliance<true>;
  * model ContosoApplianceProperties is AgentPropertiesAppliance<ContosoApplianceDefinition> {
  *   ...DefaultProvisioningStateProperty;
  * }