+++ packages/typespec-autorest/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log - @azure-tools/typespec-autorest
 
+## 0.71.0
+
+No changes, version bump only.
+
 ## 0.70.1
 
 ### Bug Fixes